:bug: 火焰图数据显示优化
